NCI-Agency/anet#255: Remove AttendeesMultiSelector component
Tiis component is not needed.
